fix: keep official DeepSeek vision CI green on Windows and repolint
Problem:
CI lint failed repolint file-size on refs.go, refs_test.go, and
anthropic.go. windows-control failed
TestResolveRefsVisionCapableImageDoesNotAskForOCR because the assertion
used filepath.Join separators.

Root cause:
Vision URL/file_id helpers landed in already-grandfathered files. The
OCR test compared an OS-native path against slash-normalized image
blocks.

Fix:
Move vision token/note helpers into vision_images.go and Anthropic
imageSourceFromRef into anthropic/image.go. Assert slash-normalized
attachment paths so Windows matches the stored ref form.

Verification:
go test ./internal/control ./internal/provider ./internal/provider/openai
./internal/provider/anthropic ./internal/provider/responses ./internal/config
File-size overages: refs.go 554/568, refs_test.go 141/141, anthropic.go 56/72.
